fix: use typescript for build-icons script
